- Nueva página para crear asientos. Añadido el modelo concepto_partida. - Nueva página de informes de artículos. Añadido el plugin jquery charts.
tpv_recambios:
 - añadido soporte de caja y borrar tickets (albaranes).
 - mejorada la búsqueda ajax de artículos, del mismo modo que en tpv_yamyam y general_paquetes.

git-svn-id: http://facturascripts.googlecode.com/svn/branches/0.9@370 fe615fe8-962f-0410-811c-9ded47c8d053

+++ model/asiento.php
@@ -26,7 +26,7 @@ class asiento extends fs_model
       {
          $this->idasiento = intval($a['idasiento']);
          $this->numero = $this->intval($a['numero']);
-         $this->idconcepto = $this->intval($a['idconcepto']);
+         $this->idconcepto = $a['idconcepto'];
          $this->concepto = $a['concepto'];
          $this->fecha = $a['fecha'];
          $this->codejercicio = $a['codejercicio'];
@@ -168,7 +168,13 @@ class asiento extends fs_model
    
    public function delete()
    {
-      return $this->db->exec("DELETE FROM ".$this->table_name." WHERE idasiento = '".$this->idasiento."';");
+      if( $this->editable )
+         return $this->db->exec("DELETE FROM ".$this->table_name." WHERE idasiento = '".$this->idasiento."';");
+      else
+      {
+         $this->new_error_msg("El asiento no es editable, por tanto no se puede borrar.");
+         return FALSE;
+      }
    }
    
    public function search($query, $offset=0)
